mark task as done

+++ src/components/ToDos/Confirm.tsx
@@ -6,17 +6,28 @@ type ConfirmProps = {
   text: string;
   taskId: string;
   confirmBtnText: string;
+  funName?: string;
   onClose?: () => void;
+  markTaskDone?: (a: string) => void;
 };
 export const Confirm = ({
   text,
   taskId,
   confirmBtnText,
   onClose,
+  markTaskDone,
+  funName,
 }: ConfirmProps) => {
   const { deleteProfileTasks } = useTasks();
   const handleSubmit = () => {
-    deleteProfileTasks(taskId);
+    if (funName === 'delete') {
+      deleteProfileTasks(taskId);
+    }
+    if (funName === 'updateTaskDone') {
+      if (markTaskDone) {
+        markTaskDone(funName);
+      }
+    }
     if (onClose) {
       onClose();
     }
